requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.18.5
openpyxl==3.0.3
pandas==1.0.4
Pillow==7.1.2
Pillow-PIL==0.1.dev0
python-dateutil==2.8.1
pytz==2020.1
scipy==1.4.1
six==1.15.0
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
